User and role

+++ database/migrations/2022_09_27_163254_create_role_user_table.php
@@ -14,9 +14,8 @@ return new class extends Migration
     public function up()
     {
         Schema::create('role_user', function (Blueprint $table) {
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->unsignedInteger('role_id');
-
             $table->foreign('user_id')->references('id')->on('users')
                 ->onDelete('cascade');
             $table->foreign('role_id')->references('id')->on('roles')
